Update code
Minor logic comment.  I processed Convert to uppercase inside the loop to check for alpha characters.  Technically it has no impact; an uppercase null string is still a null string, but to be more logically correct the conversion should only happen once a word with alpha characters is entered.  To be honest I was just so excited the program came together so well it wasn't something I thought of until review this morning.